feature(intl-messageformat): add core entry point
