+++ packages/intl-messageformat/.gitignore
@@ -2,3 +2,4 @@ lib
 tests/*.js*
 !tests/tsconfig.json
 .tsbuildinfo
+core.js
