+++ packages/intl-messageformat/rollup.config.js
@@ -27,5 +27,14 @@ export default [
     },
     plugins: [resolveConfig, uglifyConfig]
   },
+  {
+    input: './lib/core.js',
+    output: {
+      sourcemap: false,
+      file: 'core.js',
+      format: 'cjs'
+    },
+    plugins: [resolveConfig]
+  },
   ...testRollupConfig
 ];
